+++ client/src/api/index.tsx
@@ -11,6 +11,7 @@ import createRcloneAPI from './rclone';
 import createBackupsAPI from './backup';
 import createApiTokensAPI from './apiTokens';
 import createOpenIDAPI from './openid';
+import createGroupsAPI from './groups';
 
 import * as authDemo from './authentication.demo';
 import * as usersDemo from './users.demo';
@@ -55,6 +56,7 @@ export function createClient({ baseUrl, token }) {
     backups: createBackupsAPI(apiFetch),
     apiTokens: createApiTokensAPI(apiFetch),
     openid: createOpenIDAPI(apiFetch),
+    groups: createGroupsAPI(apiFetch),
 
     getStatus: () => {
       return wrap(apiFetch('/cosmos/api/status', {
@@ -390,6 +392,7 @@ let rclone = createRcloneAPI(defaultFetch);
 let backups = createBackupsAPI(defaultFetch);
 let apiTokens = createApiTokensAPI(defaultFetch);
 let openid = createOpenIDAPI(defaultFetch);
+let groups = createGroupsAPI(defaultFetch);
 
 if(isDemo) {
   auth = authDemo;
@@ -435,4 +438,5 @@ export {
   backups,
   apiTokens,
   openid,
+  groups,
 };