+++ client/src/pro/index.js
@@ -1,0 +1,6 @@
+export default {
+    isPro: false,
+    GroupsTab: null,
+    RouteLBSettings: null,
+    ConstellationNATSReplicas: null,
+}