+++ src/entreprise/index.go
@@ -1,5 +1,0 @@
-package entreprise
-
-func IsEntreprise() bool {
-	return false
-}