+++ src/pro/index.go
@@ -1,0 +1,10 @@
+package pro
+
+import "github.com/gorilla/mux"
+
+func IsPro() bool {
+	return false
+}
+
+func RegisterRoutes(router *mux.Router) {
+}